changing files names within headers, added timestamp to log files but bug with creating new log file everytime LOGGING is called. think of conditional to eliminate issue
