+++ include/remux.h
@@ -2,7 +2,7 @@
  * function declaration for our remux implementation in /src 
  */
 
-#ifndef REMUX_h
+#ifndef REMUX_H
 #define REMUX_H
 
 int remux(int argc, char **argv);
