+++ include/util.h
@@ -2,11 +2,16 @@
  * reusable util functions in one file
  */
 
-#ifndef UTIL_h
+#ifndef UTIL_H
 #define UTIL_H
 
 void rm_spc(char *str);
 
+void rep_spc(char *str);
+
+int compare_data(int a, int b);
+
+void swap_data(void* a, void* b, size_t s);
 
 #endif
